updated the text editor functionality

+++ src/components/pages/Edit.jsx
@@ -11,64 +11,67 @@ import Mantine from '../organisms/Mantine';
 import Badge from '../atoms/Badge';
 import { useEditor} from '@tiptap/react'
 import StarterKit from '@tiptap/starter-kit'
-import { Link } from '@mantine/tiptap';
-import Placeholder from '@tiptap/extension-placeholder'
-
+import Link from '@tiptap/extension-link';
+import CodeBlockLowlight from '@tiptap/extension-code-block-lowlight';
+import { createLowlight } from 'lowlight';
+import Placeholder from '@tiptap/extension-placeholder'; 
+import { IconEye, IconLock, IconPencil } from '@tabler/icons-react';
+import hljs from 'highlight.js';
+import Spinner from '../organisms/Spinner';
+import Underline from '@tiptap/extension-underline';
 function Edit() {
-  // ! lots of bugs here (will be fixed later)
+  const lowlight = createLowlight();
+ 
+  // register languages that you are planning to use
+  function escapeHtml(unsafe) {
+    return unsafe
+      .replace(/&/g, '&amp;')
+      .replace(/</g, '&lt;')
+      .replace(/>/g, '&gt;')
+      .replace(/"/g, '&quot;')
+      .replace(/'/g, '&#039;');
+  }
+
+  
+const codeExample = 'for element in name'
+
+// const highlighted = hljs.highlight(codeExample, {language: 'typescript'});
+// console.log(highlighted);
+// const content = '<p>Regular paragraph</p><pre><code>' + highlighted.value + '</code></pre>';
+// console.log(content);
+
   const note = useLoaderData();
-  const [value, setValue] = useState(note.content);
-  const [title, setTitle] = useState(note.title || 'Untitled');
+  const [saving, setSaving] = useState(false)
+  const [title, setTitle] = useState(note? note.title: 'Untitled');
   const [notes, setNotes] = useState([]);
   const quillRef = useRef();
   const titleRef = useRef();
   const editor = useEditor({
     extensions:[
-        StarterKit, 
+        StarterKit.configure({codeBlock: false}), 
         Link,
+        Underline,
+        CodeBlockLowlight.configure({ lowlight }),
         Placeholder.configure({placeholder:'Write something ...'})
     ], 
-    content: note.content ?? '', 
+    content: note? note.content:  '', 
+    editable: note? note.can_edit: true
   });
 
   useEffect(
     ()=>{
       localStorage.removeItem('current_note'); // remove existing current id's 
-      if(note.id){ // if it's an existing note then we store it's id so that we just update it
+      if(note !== null && note.id){ // if it's an existing note then we store it's id so that we just update it
         localStorage.setItem('current_note', note.id);
       }
       const func = async()=>{
-        const notes = await fetchData(DOMAIN + '/api/get-notes/', {auth: true}); 
+        if(note == null)localStorage.setItem('get_notes_for', localStorage.getItem('username'))
+        const notes = await fetchData(DOMAIN + `/api/get-notes/?username=${localStorage.getItem('get_notes_for')}`, {auth: true}); 
         if(notes!=null)setNotes(notes);
       }
       func();
     }
   , []);
-
-  const createNote = async ()=>{
-     const response = await fetch( domain + '/api/create-note/', {
-      method: 'POST', 
-      body:JSON.stringify({
-        'text': value,
-        'author': 'james',
-      }),
-      headers:{'Content-Type': 'application/json'}
-     });
-     const data = await response.json();
-     console.log(data);
-  };
-
-  const getCredentials = async ()=> {
-    const response = await fetch(domain + '/api/token/',{
-      method: 'POST',
-      body:JSON.stringify({username: 'james', password: 'easytoguess'}),
-      headers:{'Content-Type': 'application/json'}
-    });
-
-    const data = await response.json();
-    // console.log(data)
-    localStorage.setItem('user_token', data.access)
-  };
 
   const handleSave = async ()=>{
     let data;
@@ -77,24 +80,28 @@ function Edit() {
       'author': localStorage.getItem('username'), 
       'title': titleRef.current.innerText || 'Untitled',
       'brief': brief, 
-      // 'label': 'Education',
+      'label': 'Web Development',
       'content': editor.getHTML()
     }
     // if this current note has already been saved then we just perform update functions instead
+    setSaving(true);
     if(localStorage.getItem('current_note')){
       console.log('this note already exists');
       console.log(note);
       data = await fetchData(DOMAIN + `/api/update-note/${localStorage.getItem('current_note')}/` , {method:'PATCH', body: note})
       if(data != null){
         console.log(data);
+        setSaving(false);
       }
     }
     else{ // create a new note
       data = await fetchData(DOMAIN + '/api/create-note/', {method:'POST', body:note});
       if(data != null){
         localStorage.setItem('current_note', data.id)
+        setSaving(false);
       }
     }
+  
   };
 
   return (
@@ -107,7 +114,7 @@ function Edit() {
       }
 
       {notes.length > 0 && 
-          <div className="p-2 w-1/4 flex flex-col gap-y-2">
+          <div className="p-2 w-1/4 flex flex-col gap-y-2 max-h-[600px] overflow-y-auto" style={{scrollbarWidth: 'thin'}}>
             {notes.map((note) => (
               <Card key={note.id} note={note}/>
             ))}
@@ -118,7 +125,74 @@ function Edit() {
        </div> */}
      
       <div className='p-4 editor'>
-        <div className="mb-3 px-2">
+        <table className='table-auto border-separate'>
+             <tbody>
+                <tr>
+                <td>
+                  <div className='font-bold text-gray-700 me-4'>Title</div>
+                </td>
+                <td>
+                  <div contentEditable={note? note.can_edit: true} ref={titleRef} className='outline-0 font-bold'>{title}</div>
+                </td>
+              </tr>
+                
+                <tr>
+                <td>
+                 <div className='font-bold text-gray-700'>Author</div>
+                </td>
+                <td>
+                  <Badge rounded color="gray" text={note? note.author: localStorage.getItem('username')} />
+                </td>
+                </tr>
+                <tr>
+                <td>
+                 <div className='font-bold text-gray-700'>Label</div>
+                </td>
+                <td>
+                  <Badge rounded color="blue" text={note && note.label? note.label : 'empty'} />
+                </td>
+                </tr>
+             
+                <tr>
+                <td>
+                 <div className='font-bold text-gray-700'>Status</div>
+                </td>
+                <td>
+                  <div className='inline-flex items-center'>
+                  <Badge rounded color="green" text='Private' />
+                  <IconLock className='w-5 h-5'/>
+                  </div>
+                </td>
+                </tr>
+
+                <tr>
+                <td>
+                 <div className='font-bold text-gray-700'>Access</div>
+                </td>
+                <td>
+                  <div className='inline-flex items-center'>
+                  {(note == null || note.can_edit) && 
+                  <>
+                  <Badge rounded color="green" text='Read-Write' />
+                  <IconPencil className='w-5 h-5'/>
+                  </>
+                  }
+
+                  {(note != null && !note.can_edit) && 
+                  <>
+                    <Badge rounded color="orange" text='Read-Only' />
+                    <IconEye className='w-5 h-5'/>
+                  </>
+                  }
+                  </div>
+                </td>
+                </tr>
+             </tbody>
+        </table>
+        <div className={`${saving? 'opacity-1': 'opacity-0'}`}>
+           <Spinner/>
+        </div>
+        {/* <div className="mb-3 px-2">
             <div className='flex gap-3 items-center'>
               <div className='font-bold text-gray-700'>Title</div>
               <div contentEditable ref={titleRef} className='outline-0 font-bold'>{title}</div>
@@ -128,13 +202,15 @@ function Edit() {
               <div className='font-bold text-gray-700'>Label</div>
               <Badge rounded color="blue" text={note.label ?? 'empty'} />
             </div>
-        </div>
+        </div> */}
             <Mantine editor={editor}/>
+          {(note == null || note.can_edit) &&
             <div>
             <button type="button" 
             onClick={handleSave}
             className="text-white rounded-full bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium mt-4 text-sm px-5 py-2.5 me-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800">Save</button>
             </div>
+          }
        </div>
      </div>
